+++ changelog.md
@@ -1,4 +1,4 @@
-# Upcoming
+# 1.3.2.3
 * Sprint Pact speed is now configurable
 * Increased crossbow sockets
 * Projectile Protection rune now uses InsaneLib's attribute modifier